# openapi.json
{
  "components": {
    "schemas": {
...
      "Citation": {
        "description": "A pointer into a document, attached to an interjection.\n\n`quote` is the exact retrieved span. The frontend should render it verbatim \u2014\nit is the evidence, and paraphrasing it would undercut the whole feature.",
        "properties": {
          "chunk_id": {
            "examples": [
              "chk_01J8XK5B6C7D8E9F0G1H2J"
            ],
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "examples": [
              "doc_01J8XK4A5B6C7D8E9F0G1H"
            ],
            "title": "Document Id",
            "type": "string"
          },
          "filename": {
            "examples": [
              "Q3-board-deck.pdf"
            ],
            "title": "Filename",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "1-indexed. Null for non-paginated sources.",
            "title": "Page"
          },
          "quote": {
            "examples": [
              "New Product Line: $1.42M (-12.1% MoM)"
            ],
            "title": "Quote",
            "type": "string"
          },
          "relevance": {
            "examples": [
              0.91
            ],
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance",
            "type": "number"
          }
        },
        "required": [
          "document_id",
          "filename",
          "chunk_id",
          "page",
          "quote",
          "relevance"
        ],
        "title": "Citation",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code, e.g. `not_found`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context, e.g. field validation errors.",
            "title": "Detail"
          },
          "message": {
            "description": "Human-readable message. Safe to surface in UI.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Every non-2xx response from `/api` has this shape.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Interjection": {
        "properties": {
          "body_md": {
            "description": "Full reasoning, markdown. The dashboard renders this.",
            "title": "Body Md",
            "type": "string"
          },
          "chat_alert": {
            "description": "What was (or would be) posted to Meet chat. Hard-capped at 500 chars by the platform. Render verbatim so the operator sees what the meeting saw.",
            "maxLength": 500,
            "title": "Chat Alert",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "confidence": {
            "examples": [
              0.82
            ],
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "headline": {
            "examples": [
              "Sarah's revenue claim conflicts with the Q3 board deck"
            ],
            "maxLength": 200,
            "title": "Headline",
            "type": "string"
          },
          "id": {
            "examples": [
              "itj_01J8XP8Q9R0S1T2U3V4W5X"
            ],
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/InterjectionKind"
          },
          "meeting_id": {
            "title": "Meeting Id",
            "type": "string"
          },
          "posted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "spoken": {
            "default": false,
            "description": "True when Meet AGI said this out loud rather than only in chat.",
            "title": "Spoken",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/InterjectionStatus"
          },
          "trigger": {
            "$ref": "#/components/schemas/InterjectionTrigger"
          }
        },
        "required": [
          "id",
          "meeting_id",
          "kind",
          "status",
          "trigger",
          "chat_alert",
          "headline",
          "body_md",
          "confidence",
          "citations",
          "spoken",
          "error",
          "created_at",
          "posted_at"
        ],
        "title": "Interjection",
        "type": "object"
      },
      "InterjectionApprove": {
        "properties": {
          "edited_chat_alert": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override the generated alert before it posts.",
            "title": "Edited Chat Alert"
          }
        },
        "title": "InterjectionApprove",
        "type": "object"
      },
      "InterjectionDismiss": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "InterjectionDismiss",
        "type": "object"
      },
      "InterjectionKind": {
        "enum": [
          "contradiction",
          "context",
          "correction",
          "answer",
          "clarification"
        ],
        "title": "InterjectionKind",
        "type": "string"
      },
...
      "InterjectionStatus": {
        "enum": [
          "proposed",
          "approved",
          "posted",
          "dismissed",
          "failed"
        ],
        "title": "InterjectionStatus",
        "type": "string"
      },
      "InterjectionTrigger": {
        "description": "What in the meeting caused this.",
        "properties": {
          "person_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "quote": {
            "description": "The utterance that triggered this, verbatim. Rendered on the card.",
            "examples": [
              "new product revenue is up about eight percent this quarter"
            ],
            "title": "Quote",
            "type": "string"
          },
          "segment_ids": {
            "items": {
              "type": "string"
            },
            "title": "Segment Ids",
            "type": "array"
          }
        },
        "required": [
          "segment_ids",
          "person_id",
          "quote"
        ],
        "title": "InterjectionTrigger",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nBackend for **Meet AGI**, a Google Meet copilot that actually speaks.\n\n### Building the frontend against this\n\nYou do not need a Google Meet, a Recall.ai key, or an internet connection. Start a\nfixture replay and subscribe to the socket:\n\n```bash\ncurl -X POST localhost:8000/api/dev/harness/start \\\n  -H 'content-type: application/json' \\\n  -d '{\"fixture_id\":\"q3_revenue_review\",\"speed\":6,\"loop\":true}'\n```\n\nThat emits the complete real-time event stream \u2014 participants, partial and final\ntranscript, speaking states, wake-word detection, agent state transitions, and\ninterjections with real citations. It is byte-for-byte the same contract a real meeting\nproduces; only `Meeting.source` differs.\n\n### Contract rules\n\n- Types are generated from this OpenAPI document. Never hand-write an API type.\n- All mutations are REST. The WebSocket is server-to-client only.\n- Adding fields, enum variants, and event types is free and unannounced. Renames and\n  removals get a heads-up first.\n- `switch` on `LiveEvent.type` with a `default` branch. New event types will arrive.\n",
    "title": "Meet AGI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/interjections/{interjection_id}/approve": {
      "post": {
        "description": "Posts it to the meeting. Optionally overrides the generated chat alert.",
        "operationId": "approve_api_interjections__interjection_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "interjection_id",
            "required": true,
            "schema": {
              "title": "Interjection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterjectionApprove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Approve a proposed interjection",
        "tags": [
          "interjections"
        ]
      }
    },
    "/api/interjections/{interjection_id}/dismiss": {
      "post": {
        "description": "Discards it. Never reaches the meeting.",
        "operationId": "dismiss_api_interjections__interjection_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "interjection_id",
            "required": true,
            "schema": {
              "title": "Interjection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterjectionDismiss"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Dismiss a proposed interjection",
        "tags": [
          "interjections"
        ]
      }
    },
    "/api/interjections/{interjection_id}/speak": {
      "post": {
        "description": "Escalates a chat-only interjection to voice. **Milestone 7** performs the actual TTS; this marks `spoken` and drives the agent state so the UI can be built.",
        "operationId": "speak_api_interjections__interjection_id__speak_post",
        "parameters": [
          {
            "in": "path",
            "name": "interjection_id",
            "required": true,
            "schema": {
              "title": "Interjection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Have Meet AGI say an interjection out loud",
        "tags": [
          "interjections"
        ]
      }
    },
...
  }
}